LibWeb: Split GridTrackPlacementStyleValue out of StyleValue.{h,cpp}
