+++ Userland/Libraries/LibWeb/CSS/StyleValues/GridAreaShorthandStyleValue.cpp
@@ -8,6 +8,7 @@
  */
 
 #include "GridAreaShorthandStyleValue.h"
+#include <LibWeb/CSS/StyleValues/GridTrackPlacementStyleValue.h>
 
 namespace Web::CSS {
